+++ types/index.d.ts
@@ -1,0 +1,3 @@
+import * as api from './pm-rpc';
+
+export { api };